refactor: fix workflow, change maps to objects, rename OIDC file
