+++ .github/workflows/e2e-jimm-k8s-charm-oidc-auth.yml
@@ -52,7 +52,7 @@ jobs:
         with:
           charm-channel: ${{ steps.config.outputs.dashboard-charm-channel }}
           controller-app: jimm
-          controller-model: '${{ steps.jimm.outputs.jimm-controller-name }}':jimm
+          controller-model: '${{ steps.jimm.outputs.jimm-controller-name }}:jimm'
           dashboard-charm-ref: ${{ steps.config.outputs.dashboard-charm-ref }}
           dashboard-ref: ${{ steps.config.outputs.repo-ref }}
           dashboard-resource: ${{ steps.config.outputs.k8s-dashboard-resource }}
