+++ e2e/helpers/auth/backends/index.ts
@@ -1,2 +1,3 @@
 export { CreateLocalUser, LocalUser } from "./Local";
 export { CreateCandidUser, CandidUser } from "./Candid";
+export { CreateOIDCUser, OIDCUser } from "./OIDC";
